+++ include/game/player.hpp
@@ -12,15 +12,15 @@ namespace Game
 
     public:
         Player(const Math::Point2 &position, const Math::Point2 &lookAt,
-               double fov, double focalDistance, unsigned resolution);
+               float fov, float focalDistance, unsigned resolution);
         ~Player() = default;
 
         const Math::Point2 &getPosition() const;
         const Game::Camera &getCamera() const;
 
-        void update(double dt);
+        void update(float dt);
 
         void move(const Math::Vector2 &delta);
-        void rotate(double angle);
+        void rotate(float angle);
     };
 } // namespace Game
